+++ app/src/main/java/com/rodrigotristany/mrjeff/data/weather/models/WeatherRequest.kt
@@ -1,8 +1,8 @@
 package com.rodrigotristany.mrjeff.data.weather.models
 
 class WeatherRequest(
-    val north: Float,
-    val south: Float,
-    val east: Float,
-    val west: Float
+    val north: Double,
+    val south: Double,
+    val east: Double,
+    val west: Double
 )
